Factor out `Dotpkg` and `DotpkgRef` into model module
